+++ client-modern/src/main/java/com/paraguacraft/pvp/modern/ParaguacraftPvPModern.java
@@ -23,7 +23,7 @@ import net.minecraft.client.MinecraftClient;
 public class ParaguacraftPvPModern implements ClientModInitializer {
 
     public static final String MOD_ID = "paraguacraftpvp-modern";
-    public static final String VERSION = "0.9.2";
+    public static final String VERSION = "0.9.3";
 
     @Override
     public void onInitializeClient() {